Try reg <- constant loading during address mod generation.
This makes the generator notice cases when e.g. loading the constant
displacement of an address into the index reg and doing a @(base+index)
access is cheaper than offseting the base reg.
With this change, it's no longer neccessary to add (reg <- constant)
reg_mod accesses during insn parsing, because they're added by the
reg mod generator if neccessary.