Use a more efficient way to generate address modifications.
Store the "starting addresses" (i.e. the addresses that are used as a
starting point when trying to arrive at the required addresses) in a way
that allows fast retrieval of the relevant start addresses for a given
end address.  This way the address mod generator doesn't have to go over
the entire access sequence for every mem-access/reg-use/etc. that it
tries to optimize.