Add a delegate function to adjust alternative costs.
The new adjust_alternative_costs function is called on each alternative
of the memory accesses in the sequence.  Delegates can use the function
along with alternative::update_costs to adjust the cost of the
alternatives based on the surrounding accesses.

Use adjust_alternative_costs and lookahead_count to make AMS use the
post-inc alternatives instead of base+disp for adjacent QImode/HImode
accesses.